requirements.txt
git+https://github.com/huggingface/transformers
datasets 
loralib 
sentencepiece
bitsandbytes 
accelerate 
xformers 
einops 
langchain 
streamlit 
peft
ffmpeg-python 
pydub
streamlit-audiorecorder